Fixed ingrdients autocomplete and recipe visualisation

+++ app/schema.sql
@@ -41,7 +41,7 @@ CREATE TABLE ingredients (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
   recipe_id INTEGER NOT NULL,
   ingredient_id INTEGER NOT NULL,
-  quantity INTEGER NOT NULL,
+  quantity REAL NOT NULL,
   unit TEXT NOT NULL, -- g, ml, pieces of, teaspoons
   FOREIGN KEY (recipe_id) REFERENCES recipes (id),
   FOREIGN KEY (ingredient_id) REFERENCES ingredient_type (id)
